Further simplifying editor registration and resolution
