+++ metatonic-core/src/services/EditorContext.ts
@@ -1,82 +1,15 @@
-import {RecordSchemaType, SchemaField, SchemaRecordTypeParameters} from "../domain/Schema/Records";
+import {RecordSchemaType, SchemaField} from "../domain/Schema/Records";
 import {SchemaEntryType} from "../domain/Schema/SchemaEnums";
 import {ItemCollectionSize} from "../domain/Schema/ItemSelectionType";
 import {FormSchema} from "../domain/Schema/RootSchemas";
-
-type LabeledEditor<TEditor, TLabeler, TRepeater> = { editor: TEditor, labeler: TLabeler, repeater?: TRepeater };
-type TypeRegistration<TEditor, TLabeler, TRepeater> = {
-    availableComponents: LabeledEditor<TEditor, TLabeler, TRepeater>[];
-    uiHintMap: { [uiHint: string]: LabeledEditor<TEditor, TLabeler, TRepeater> };
-    defaultComponent?: LabeledEditor<TEditor, TLabeler, TRepeater>;
-}
-type TypeEditorRegistry<TEditor, TLabeler, TRepeater> = {
-    [type:string]: TypeRegistration<TEditor, TLabeler, TRepeater>
-}
-
-export class EditorRegistrationSubContext<
-    TEditor extends new (...args) => any,
-    TLabeler extends new (...args) => any,
-    TRepeater extends new (...args) => any> {
-    editorRegistrations: TypeEditorRegistry<TEditor, TLabeler, TRepeater>;
-    repeater: TRepeater;
-
-    constructor(protected schema: FormSchema, protected select: boolean) {
-    }
-
-    registerComponent<TData, TType, TParams, TState>(
-        type: string,
-        editor: TEditor,
-        labeler: TLabeler,
-        options?: {
-            uiHint?: string | string[],
-            isDefault?: boolean,
-            repeater?: TRepeater
-        }
-    ) {
-        options = options || {}
-        let uiHint = options.uiHint;
-        this.createNewTypeEntryIfNeeded(type);
-        let labeledEditor = { editor, labeler, repeater: options.repeater };
-        let typeEntry = this.editorRegistrations[type];
-        typeEntry.availableComponents.push(labeledEditor);
-
-        if (uiHint) {
-            if (Array.isArray(uiHint))
-                uiHint.forEach(hint => typeEntry[hint] = labeledEditor);
-            else
-                typeEntry[uiHint] = labeledEditor;
-        }
-
-        if (options.isDefault) {
-            typeEntry.defaultComponent = labeledEditor;
-        }
-    }
-
-    defaultRepeater(repeater: TRepeater) {
-        this.repeater = repeater;
-    }
-
-    createNewTypeEntryIfNeeded(type: string) {
-        if (!this.editorRegistrations[type]) {
-            this.editorRegistrations[type] = this.emptyTypeEntry()
-        }
-    }
-
-    emptyTypeEntry() {
-        return {
-            availableComponents: [],
-            defaultComponent: undefined,
-            uiHintMap: {}
-        } as TypeRegistration<TEditor, TLabeler, TRepeater>;
-    }
-}
+import {EditorRegistry} from "./EditorRegistry";
 
 export class EditorSubContext<
     TEditor extends new (...args) => any,
     TLabeler extends new (...args) => any,
     TRepeater extends new (...args) => any> {
 
-    constructor(protected editors: EditorRegistrationSubContext<TEditor, TLabeler, TRepeater>,
+    constructor(protected editors: EditorRegistry<TEditor, TLabeler, TRepeater>,
                 protected schema: FormSchema,
                 protected select: boolean) {
 
@@ -140,8 +73,8 @@ export class EditorContext<
         }
     }
 }
-export var editorConfig: {
-    mainEditorRegistrationContext?: EditorRegistrationSubContext<any, any, any>
-    mainMultiEditorRegistrationContext?: EditorRegistrationSubContext<any, any, any>
-    mainSelectRegistrationContext?: EditorRegistrationSubContext<any, any, any>
-}
+
+
+
+
+
