+++ metatonic-core/src/services/index.ts
@@ -1,5 +1,5 @@
 export * from './PropertySelection'
-export * from './BaseEditorService'
+export * from './EditorContext'
 export * from './ContextService'
 export * from './HtmlInputTypeService'
 export * from './IdGeneratorService'
@@ -7,3 +7,4 @@ export * from './NumericMaskService'
 export * from './QuantityFieldService'
 export * from './Rest'
 export * from './SchemaFromJsonService'
+export * from "./EditorRegistry";
